refactor: transfer positionsMap const to a sepaat modulere
