+++ roulette/models/str_upModel.js
@@ -22,14 +22,14 @@ const positions = document.getElementById('positions');
 const bet = document.getElementById('bet');
 const color = document.getElementById('color');
 const cash = document.getElementById('cash');
+const info = document.getElementById('info_3');
 
 let fillArr = [];
 
 // клик по кнопке Ставка ОК (подсчёт кол-ва фишек):
 add_bet.onclick = function () {
-    fillArr = chipCount(positions, bet); // Вызываем chipCount и сохраняем результат в fillArr
+    fillArr = chipCount(positions, bet); // сохраняем результат в fillArr
     console.log(fillArr);
-    console.log(`positions: ${positions.value}`);
 };
 
 // клик по кнопке Рассчитать:
